+++ src/rust/engine/process_execution/docker/Cargo.toml
@@ -21,14 +21,9 @@ tokio = { version = "1.21", features = ["net", "process", "rt-multi-thread", "sy
 tokio-rustls = "0.23"
 tokio-util = { version = "0.7", features = ["codec"] }
 workunit_store = { path = "../../workunit_store" }
-lazy_static = "1"
 parking_lot = "0.12"
 async-oncecell = "0.2"
 once_cell = "1.15"
-os_pipe = "1.0"
-strum = "0.24"
-strum_macros = "0.24"
-tryfuture = { path = "../../tryfuture" }
 process_execution = { path = ".." }
 
 [dev-dependencies]
