+++ src/rust/engine/process_execution/pe_nailgun/src/lib.rs
@@ -1,5 +1,31 @@
 // Copyright 2022 Pants project contributors (see CONTRIBUTORS.md).
 // Licensed under the Apache License, Version 2.0 (see LICENSE).
+//
+//
+#![deny(warnings)]
+// Enable all clippy lints except for many of the pedantic ones. It's a shame this needs to be copied and pasted across crates, but there doesn't appear to be a way to include inner attributes from a common source.
+#![deny(
+  clippy::all,
+  clippy::default_trait_access,
+  clippy::expl_impl_clone_on_copy,
+  clippy::if_not_else,
+  clippy::needless_continue,
+  clippy::unseparated_literal_suffix,
+  clippy::used_underscore_binding
+)]
+// It is often more clear to show that nothing is being moved.
+#![allow(clippy::match_ref_pats)]
+// Subjective style.
+#![allow(
+  clippy::len_without_is_empty,
+  clippy::redundant_field_names,
+  clippy::too_many_arguments
+)]
+// Default isn't as big a deal as people seem to think it is.
+#![allow(clippy::new_without_default, clippy::new_ret_no_self)]
+// Arc<Mutex> can be more clear than needing to grok Orderings:
+#![allow(clippy::mutex_atomic)]
+
 use std::collections::{BTreeMap, BTreeSet};
 use std::fmt::{self, Debug};
 use std::net::SocketAddr;
@@ -15,8 +41,8 @@ use task_executor::Executor;
 use tokio::net::TcpStream;
 use workunit_store::{in_workunit, Metric, RunningWorkunit};
 
-use crate::local::{prepare_workdir, CapturedWorkdir, ChildOutput};
-use crate::{
+use process_execution::local::{prepare_workdir, CapturedWorkdir, ChildOutput};
+use process_execution::{
   Context, FallibleProcessResultWithPlatform, InputDigests, NamedCaches, Platform, Process,
   ProcessError,
 };
@@ -131,7 +157,7 @@ impl Debug for CommandRunner {
 }
 
 #[async_trait]
-impl super::CommandRunner for CommandRunner {
+impl process_execution::CommandRunner for CommandRunner {
   async fn run(
     &self,
     context: Context,
