+++ src/rust/engine/process_execution/pe_nailgun/src/parsed_jvm_command_lines_tests.rs
@@ -1,6 +1,6 @@
 // Copyright 2022 Pants project contributors (see CONTRIBUTORS.md).
 // Licensed under the Apache License, Version 2.0 (see LICENSE).
-use crate::nailgun::parsed_jvm_command_lines::ParsedJVMCommandLines;
+use crate::parsed_jvm_command_lines::ParsedJVMCommandLines;
 
 // TODO we should be able to use https://docs.rs/crate/derive_builder/0.8.0
 #[derive(Debug)]
