+++ src/rust/engine/process_executor/Cargo.toml
@@ -16,7 +16,6 @@ grpc_util = { path = "../grpc_util" }
 hashing = { path = "../hashing" }
 log = "0.4"
 process_execution = { path = "../process_execution" }
-docker = { path = "../process_execution/docker"}
 prost = "0.9"
 shlex = "1.1.0"
 store = { path = "../fs/store" }
